Start of a parser for data from the new wikiwho beta API
This is intended to replace the whoColor server currently used for consuming article markup and wikiwho data, and returning article markup with added authorship annotations (used for the authorship highlighting feature of the ArticleViewer).

The server would require updates to work with the new wikiwho data format. Implementing it in Javascript instead would mean that we can eliminate the need for a whoColor server service altogether, and just consume wikiwho directly from the client.

This initial version is incomplete and does not handle the full variety of markup that whoColor PHP implementation does, but it works for the simple article 'Honors at Dawn'.

See https://github.com/jurkan/whoCOLOR/blob/master/server/markuppreparser.inc.php